+++ catTown/Assets/Hyeon/Material/Mat_ground.mat
@@ -78,6 +78,6 @@ Material:
     - _UVSec: 0
     - _ZWrite: 1
     m_Colors:
-    - _Color: {r: 0.81761, g: 0.7481902, b: 0.8007343, a: 1}
+    - _Color: {r: 0.2578616, g: 0.21001935, b: 0.24660462, a: 1}
     - _EmissionColor: {r: 0, g: 0, b: 0, a: 1}
   m_BuildTextureStacks: []
