--- catTown/Assets/Hyeon/Script/AttackScript.cs
@@ -1,24 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class AttackScript : MonoBehaviour
-{
-
-    private void  OnCollicisionEnter(Collision collision)
-    {
-
-        Destroy(gameObject);
-    }
-    // Start is called before the first frame update
-    void Start()
-    {
-        
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-        
-    }
-}
