+++ catTown/Assets/Hyeon/Script/CamFollow.cs
@@ -1,0 +1,18 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class CamFollow : MonoBehaviour
+{
+    public Transform target;
+    
+    void Start()
+    {
+        
+    }
+
+    void Update()
+    {
+        transform.position = target.position;
+    }
+}
