+++ catTown/Assets/Hyeon/Script/CamFollow.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: e599d3aa536c3d74f937add48aa3cc57
+guid: d08e392fa07b2fa4ca554599ab35ca11
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
