+++ catTown/Assets/Hyeon/Script/CamRotate.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 54a3a45c3ff577242bade93a0b99ef1e
+guid: 6b961c0dfda25fd4584381f4e3ce9d93
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
