--- catTown/Assets/Hyeon/player.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 8db0e98197be5bf489e714e0c6130da4
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
